requirements.txt
fastapi
uvicorn
python-dotenv
pydantic

qdrant-client==1.16.2
httpx

google-genai
pypdf
python-multipart